requirements.txt
-i https://pypi.org/simple
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
docker==4.2.0
flask-restx==0.2.0
flask-sqlalchemy==2.4.1
flask==0.12.2
gitdb==4.0.5
gitpython==3.1.2
humanfriendly==8.2
idna==2.9
importlib-metadata==1.6.0 ; python_version < '3.8'
itsdangerous==1.1.0
jinja2==2.11.2
jsonschema==3.2.0
markupsafe==1.1.1
pypiwin32==223 ; sys_platform == 'win32' and python_version >= '3.6'
pyreadline==2.1 ; sys_platform == 'win32'
pyrsistent==0.16.0
pytz==2020.1
requests==2.23.0
six==1.15.0
smmap==3.0.4
sqlalchemy==1.3.17
urllib3==1.25.9
websocket-client==0.57.0
werkzeug==0.16.1
wtforms==2.3.1
zipp==3.1.0
gunicorn==20.0.4

